feat(validators): rename duration to include iso8601 prefixes
