+++ src/index.js
@@ -86,9 +86,9 @@ import url from './validators/url';
 import urlProtocol from './validators/url-protocol';
 import uuid from './validators/uuid';
 import mustHaveAllPrefixes from './validators/must-have-all-prefixes';
-import duration from './validators/duration';
-import minDuration from './validators/min-duration';
-import maxDuration from './validators/max-duration';
+import iso8601Duration from './validators/iso-8601-duration';
+import iso8601MinDuration from './validators/iso-8601-min-duration';
+import iso8601MaxDuration from './validators/iso-8601-max-duration';
 
 let validators = [
   alpha,
@@ -113,7 +113,6 @@ let validators = [
   dateTimeBefore,
   dateTimeBeforeOrEqual,
   dateFormat,
-  duration,
   email,
   emptyString,
   equal,
@@ -131,12 +130,13 @@ let validators = [
   integer,
   internationalPhoneNumber,
   ip,
+  iso8601Duration,
+  iso8601MinDuration,
+  iso8601MaxDuration,
   length,
-  maxDuration,
   maxLength,
   maxValue,
   memberOf,
-  minDuration,
   minLength,
   minValue,
   minimumAge,
